test-requirements.txt
pytest
pytest-cov
pytest-xdist
mock
python-magic
